requirements.txt
torch>=2.1.0
torchvision
pillow
albumentations
tqdm
pandas
jittor==1.3.9.14
numpy==1.26.4